Made WidgetAttraction a Util class;
