+++ src/main/java/com/mrcrayfish/controllable/client/ControllerInput.java
@@ -13,6 +13,8 @@ import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.screen.inventory.ContainerScreen;
 import net.minecraft.client.gui.screen.inventory.CreativeScreen;
 import net.minecraft.client.gui.screen.inventory.InventoryScreen;
+import net.minecraft.client.gui.widget.AbstractSlider;
+import net.minecraft.client.gui.widget.Widget;
 import net.minecraft.client.util.NativeUtil;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.inventory.container.Slot;
@@ -37,6 +39,7 @@ import org.lwjgl.glfw.GLFW;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.util.List;
 
 import static org.libsdl.SDL.SDL_CONTROLLER_BUTTON_DPAD_DOWN;
 import static org.libsdl.SDL.SDL_CONTROLLER_BUTTON_DPAD_UP;
@@ -68,7 +71,8 @@ public class ControllerInput
 
     private int dropCounter = -1;
 
-    private WidgetAttraction widgetAttraction = new WidgetAttraction(this);
+    private Screen prevScreen;
+    private List<Widget> widgetList;
 
     public double getVirtualMouseX()
     {
@@ -173,7 +177,7 @@ public class ControllerInput
             prevXAxis = controller.getLThumbStickXValue();
             prevYAxis = controller.getLThumbStickYValue();
 
-            this.widgetAttraction.moveMouseToClosestWidget(moving, mc.currentScreen);
+            this.moveMouseToClosestWidget(moving, mc.currentScreen);
             this.moveMouseToClosestSlot(moving, mc.currentScreen);
 
             if(mc.currentScreen instanceof CreativeScreen)
@@ -642,6 +646,61 @@ public class ControllerInput
         catch(IllegalAccessException | InvocationTargetException e)
         {
             e.printStackTrace();
+        }
+    }
+
+
+    public void moveMouseToClosestWidget(boolean moving, Screen screen)
+    {
+        Minecraft mc = Minecraft.getInstance();
+
+        if (screen != prevScreen)
+        {
+            widgetList.clear();
+        }
+        if(widgetList.isEmpty() && screen != prevScreen)
+        {
+            widgetList = WidgetAttraction.getWidgets(screen);
+            prevScreen = screen;
+            if(widgetList.isEmpty())
+            {
+                return;
+            }
+        }
+
+        if(!this.moved) return;
+
+        int mouseX = (int) (targetMouseX * (double) mc.mainWindow.getScaledWidth() / (double) mc.mainWindow.getWidth());
+        int mouseY = (int) (targetMouseY * (double) mc.mainWindow.getScaledHeight() / (double) mc.mainWindow.getHeight());
+
+        Widget closestWidget = null;
+        for(Widget widget : widgetList)
+        {
+            if (widget == null)
+            {
+                continue;
+            }
+
+            if(widget.isHovered() && widget.visible)
+            {
+                closestWidget = widget;
+            }
+        }
+
+        if(closestWidget != null)
+        {
+            int targetX = closestWidget.x + closestWidget.getWidth() / 2;
+            if (closestWidget instanceof AbstractSlider)
+            {
+                targetX = mouseX;
+            }
+            int targetY = closestWidget.y + closestWidget.getHeight() / 2;
+            moveMouse(mc, moving, targetX, targetY, mouseX, mouseY);
+        }
+        else
+        {
+            mouseSpeedX *= 0.1F;
+            mouseSpeedY *= 0.1F;
         }
     }
